@@ -1,7 +1,10 @@
-// Autogenerated from Pigeon (v10.1.6), do not edit directly.
+// Autogenerated from Pigeon (v16.0.0), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 
 package dr.achim.sleep_timer;
+
+import static java.lang.annotation.ElementType.METHOD;
+import static java.lang.annotation.RetentionPolicy.CLASS;
 
 import android.util.Log;
 import androidx.annotation.NonNull;
@@ -11,6 +14,8 @@ import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MessageCodec;
 import io.flutter.plugin.common.StandardMessageCodec;
 import java.io.ByteArrayOutputStream;
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -56,6 +61,15 @@ public class Messages {
     }
     return errorList;
   }
+
+  @NonNull
+  protected static FlutterError createConnectionError(@NonNull String channelName) {
+    return new FlutterError("channel-error",  "Unable to establish connection on channel: " + channelName + ".", "");
+  }
+
+  @Target(METHOD)
+  @Retention(CLASS)
+  @interface CanIgnoreReturnValue {}
 
   /** Generated class from Pigeon that represents data sent in messages. */
   public static final class NotificationRequest {
@@ -153,6 +167,7 @@ public class Messages {
 
       private @Nullable String timerId;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTimerId(@Nullable String setterArg) {
         this.timerId = setterArg;
         return this;
@@ -160,6 +175,7 @@ public class Messages {
 
       private @Nullable String title;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTitle(@Nullable String setterArg) {
         this.title = setterArg;
         return this;
@@ -167,6 +183,7 @@ public class Messages {
 
       private @Nullable String description;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setDescription(@Nullable String setterArg) {
         this.description = setterArg;
         return this;
@@ -174,6 +191,7 @@ public class Messages {
 
       private @Nullable Long accentColor;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setAccentColor(@Nullable Long setterArg) {
         this.accentColor = setterArg;
         return this;
@@ -181,6 +199,7 @@ public class Messages {
 
       private @Nullable String restartAction;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setRestartAction(@Nullable String setterArg) {
         this.restartAction = setterArg;
         return this;
@@ -188,6 +207,7 @@ public class Messages {
 
       private @Nullable String continueAction;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setContinueAction(@Nullable String setterArg) {
         this.continueAction = setterArg;
         return this;
@@ -195,6 +215,7 @@ public class Messages {
 
       private @Nullable String pauseAction;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setPauseAction(@Nullable String setterArg) {
         this.pauseAction = setterArg;
         return this;
@@ -202,6 +223,7 @@ public class Messages {
 
       private @Nullable String cancelAction;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setCancelAction(@Nullable String setterArg) {
         this.cancelAction = setterArg;
         return this;
@@ -209,6 +231,7 @@ public class Messages {
 
       private @Nullable List<Long> extendActions;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setExtendActions(@Nullable List<Long> setterArg) {
         this.extendActions = setterArg;
         return this;
@@ -386,6 +409,7 @@ public class Messages {
 
       private @Nullable String description;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setDescription(@Nullable String setterArg) {
         this.description = setterArg;
         return this;
@@ -393,6 +417,7 @@ public class Messages {
 
       private @Nullable String timerId;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTimerId(@Nullable String setterArg) {
         this.timerId = setterArg;
         return this;
@@ -400,6 +425,7 @@ public class Messages {
 
       private @Nullable String title;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTitle(@Nullable String setterArg) {
         this.title = setterArg;
         return this;
@@ -407,6 +433,7 @@ public class Messages {
 
       private @Nullable Long accentColor;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setAccentColor(@Nullable Long setterArg) {
         this.accentColor = setterArg;
         return this;
@@ -414,6 +441,7 @@ public class Messages {
 
       private @Nullable String restartAction;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setRestartAction(@Nullable String setterArg) {
         this.restartAction = setterArg;
         return this;
@@ -421,6 +449,7 @@ public class Messages {
 
       private @Nullable String continueAction;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setContinueAction(@Nullable String setterArg) {
         this.continueAction = setterArg;
         return this;
@@ -428,6 +457,7 @@ public class Messages {
 
       private @Nullable String pauseAction;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setPauseAction(@Nullable String setterArg) {
         this.pauseAction = setterArg;
         return this;
@@ -435,6 +465,7 @@ public class Messages {
 
       private @Nullable String cancelAction;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setCancelAction(@Nullable String setterArg) {
         this.cancelAction = setterArg;
         return this;
@@ -442,6 +473,7 @@ public class Messages {
 
       private @Nullable List<Long> extendActions;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setExtendActions(@Nullable List<Long> setterArg) {
         this.extendActions = setterArg;
         return this;
@@ -449,6 +481,7 @@ public class Messages {
 
       private @Nullable Long duration;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setDuration(@Nullable Long setterArg) {
         this.duration = setterArg;
         return this;
@@ -456,6 +489,7 @@ public class Messages {
 
       private @Nullable Long remainingTime;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setRemainingTime(@Nullable Long setterArg) {
         this.remainingTime = setterArg;
         return this;
@@ -650,6 +684,7 @@ public class Messages {
 
       private @Nullable String description;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setDescription(@Nullable String setterArg) {
         this.description = setterArg;
         return this;
@@ -657,6 +692,7 @@ public class Messages {
 
       private @Nullable String timerId;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTimerId(@Nullable String setterArg) {
         this.timerId = setterArg;
         return this;
@@ -664,6 +700,7 @@ public class Messages {
 
       private @Nullable String title;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTitle(@Nullable String setterArg) {
         this.title = setterArg;
         return this;
@@ -671,6 +708,7 @@ public class Messages {
 
       private @Nullable Long accentColor;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setAccentColor(@Nullable Long setterArg) {
         this.accentColor = setterArg;
         return this;
@@ -678,6 +716,7 @@ public class Messages {
 
       private @Nullable String restartAction;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setRestartAction(@Nullable String setterArg) {
         this.restartAction = setterArg;
         return this;
@@ -685,6 +724,7 @@ public class Messages {
 
       private @Nullable String continueAction;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setContinueAction(@Nullable String setterArg) {
         this.continueAction = setterArg;
         return this;
@@ -692,6 +732,7 @@ public class Messages {
 
       private @Nullable String pauseAction;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setPauseAction(@Nullable String setterArg) {
         this.pauseAction = setterArg;
         return this;
@@ -699,6 +740,7 @@ public class Messages {
 
       private @Nullable String cancelAction;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setCancelAction(@Nullable String setterArg) {
         this.cancelAction = setterArg;
         return this;
@@ -706,6 +748,7 @@ public class Messages {
 
       private @Nullable List<Long> extendActions;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setExtendActions(@Nullable List<Long> setterArg) {
         this.extendActions = setterArg;
         return this;
@@ -713,6 +756,7 @@ public class Messages {
 
       private @Nullable Long duration;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setDuration(@Nullable Long setterArg) {
         this.duration = setterArg;
         return this;
@@ -720,6 +764,7 @@ public class Messages {
 
       private @Nullable Long remainingTime;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setRemainingTime(@Nullable Long setterArg) {
         this.remainingTime = setterArg;
         return this;
@@ -727,6 +772,7 @@ public class Messages {
 
       private @Nullable Boolean shakeToExtend;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setShakeToExtend(@Nullable Boolean setterArg) {
         this.shakeToExtend = setterArg;
         return this;
@@ -824,6 +870,7 @@ public class Messages {
 
       private @Nullable String timerId;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTimerId(@Nullable String setterArg) {
         this.timerId = setterArg;
         return this;
@@ -831,6 +878,7 @@ public class Messages {
 
       private @Nullable Boolean success;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setSuccess(@Nullable Boolean setterArg) {
         this.success = setterArg;
         return this;
@@ -878,6 +926,7 @@ public class Messages {
 
       private @Nullable String timerId;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTimerId(@Nullable String setterArg) {
         this.timerId = setterArg;
         return this;
@@ -931,6 +980,7 @@ public class Messages {
 
       private @Nullable String timerId;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTimerId(@Nullable String setterArg) {
         this.timerId = setterArg;
         return this;
@@ -938,6 +988,7 @@ public class Messages {
 
       private @Nullable Boolean success;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setSuccess(@Nullable Boolean setterArg) {
         this.success = setterArg;
         return this;
@@ -996,6 +1047,7 @@ public class Messages {
 
       private @Nullable String timerId;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTimerId(@Nullable String setterArg) {
         this.timerId = setterArg;
         return this;
@@ -1003,6 +1055,7 @@ public class Messages {
 
       private @Nullable Long additionalTime;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setAdditionalTime(@Nullable Long setterArg) {
         this.additionalTime = setterArg;
         return this;
@@ -1061,6 +1114,7 @@ public class Messages {
 
       private @Nullable String timerId;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTimerId(@Nullable String setterArg) {
         this.timerId = setterArg;
         return this;
@@ -1068,6 +1122,7 @@ public class Messages {
 
       private @Nullable Long newTime;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setNewTime(@Nullable Long setterArg) {
         this.newTime = setterArg;
         return this;
@@ -1115,6 +1170,7 @@ public class Messages {
 
       private @Nullable String timerId;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTimerId(@Nullable String setterArg) {
         this.timerId = setterArg;
         return this;
@@ -1158,6 +1214,7 @@ public class Messages {
 
       private @Nullable String timerId;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTimerId(@Nullable String setterArg) {
         this.timerId = setterArg;
         return this;
@@ -1201,6 +1258,7 @@ public class Messages {
 
       private @Nullable String title;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTitle(@Nullable String setterArg) {
         this.title = setterArg;
         return this;
@@ -1264,6 +1322,7 @@ public class Messages {
 
       private @Nullable String title;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setTitle(@Nullable String setterArg) {
         this.title = setterArg;
         return this;
@@ -1271,6 +1330,7 @@ public class Messages {
 
       private @Nullable String icon;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setIcon(@Nullable String setterArg) {
         this.icon = setterArg;
         return this;
@@ -1278,6 +1338,7 @@ public class Messages {
 
       private @Nullable String packageName;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setPackageName(@Nullable String setterArg) {
         this.packageName = setterArg;
         return this;
@@ -1329,6 +1390,7 @@ public class Messages {
 
       private @Nullable List<Package> apps;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setApps(@Nullable List<Package> setterArg) {
         this.apps = setterArg;
         return this;
@@ -1372,6 +1434,7 @@ public class Messages {
 
       private @Nullable String packageName;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setPackageName(@Nullable String setterArg) {
         this.packageName = setterArg;
         return this;
@@ -1415,6 +1478,7 @@ public class Messages {
 
       private @Nullable Boolean enable;
 
+      @CanIgnoreReturnValue
       public @NonNull Builder setEnable(@Nullable Boolean setterArg) {
         this.enable = setterArg;
         return this;
@@ -1440,6 +1504,31 @@ public class Messages {
       pigeonResult.setEnable((Boolean) enable);
       return pigeonResult;
     }
+  }
+
+  /** Asynchronous error handling return type for non-nullable API method returns. */
+  public interface Result<T> {
+    /** Success case callback method for handling returns. */
+    void success(@NonNull T result);
+
+    /** Failure case callback method for handling errors. */
+    void error(@NonNull Throwable error);
+  }
+  /** Asynchronous error handling return type for nullable API method returns. */
+  public interface NullableResult<T> {
+    /** Success case callback method for handling returns. */
+    void success(@Nullable T result);
+
+    /** Failure case callback method for handling errors. */
+    void error(@NonNull Throwable error);
+  }
+  /** Asynchronous error handling return type for void API method returns. */
+  public interface VoidResult {
+    /** Success case callback method for handling returns. */
+    void success();
+
+    /** Failure case callback method for handling errors. */
+    void error(@NonNull Throwable error);
   }
 
   private static class HostTimerApiCodec extends StandardMessageCodec {
@@ -1543,7 +1632,7 @@ public class Messages {
       return HostTimerApiCodec.INSTANCE;
     }
     /**Sets up an instance of `HostTimerApi` to handle messages through the `binaryMessenger`. */
-    static void setup(@NonNull BinaryMessenger binaryMessenger, @Nullable HostTimerApi api) {
+    static void setUp(@NonNull BinaryMessenger binaryMessenger, @Nullable HostTimerApi api) {
       {
         BasicMessageChannel<Object> channel =
             new BasicMessageChannel<>(
@@ -1790,97 +1879,213 @@ public class Messages {
     }
 
     /** Public interface for sending reply. */ 
-    @SuppressWarnings("UnknownNullness")
-    public interface Reply<T> {
-      void reply(T reply);
-    }
     /** The codec used by FlutterTimerApi. */
     static @NonNull MessageCodec<Object> getCodec() {
       return FlutterTimerApiCodec.INSTANCE;
     }
-    public void onExtendTime(@NonNull ExtendTimeRequest requestArg, @NonNull Reply<Void> callback) {
+    public void onExtendTime(@NonNull ExtendTimeRequest requestArg, @NonNull VoidResult result) {
+      final String channelName = "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onExtendTime";
       BasicMessageChannel<Object> channel =
           new BasicMessageChannel<>(
-              binaryMessenger, "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onExtendTime", getCodec());
+              binaryMessenger, channelName, getCodec());
       channel.send(
           new ArrayList<Object>(Collections.singletonList(requestArg)),
-          channelReply -> callback.reply(null));
+          channelReply -> {
+            if (channelReply instanceof List) {
+              List<Object> listReply = (List<Object>) channelReply;
+              if (listReply.size() > 1) {
+                result.error(new FlutterError((String) listReply.get(0), (String) listReply.get(1), (String) listReply.get(2)));
+              } else {
+                result.success();
+              }
+            }  else {
+              result.error(createConnectionError(channelName));
+            } 
+          });
     }
-    public void onCountDown(@NonNull CountDownRequest requestArg, @NonNull Reply<Void> callback) {
+    public void onCountDown(@NonNull CountDownRequest requestArg, @NonNull VoidResult result) {
+      final String channelName = "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onCountDown";
       BasicMessageChannel<Object> channel =
           new BasicMessageChannel<>(
-              binaryMessenger, "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onCountDown", getCodec());
+              binaryMessenger, channelName, getCodec());
       channel.send(
           new ArrayList<Object>(Collections.singletonList(requestArg)),
-          channelReply -> callback.reply(null));
+          channelReply -> {
+            if (channelReply instanceof List) {
+              List<Object> listReply = (List<Object>) channelReply;
+              if (listReply.size() > 1) {
+                result.error(new FlutterError((String) listReply.get(0), (String) listReply.get(1), (String) listReply.get(2)));
+              } else {
+                result.success();
+              }
+            }  else {
+              result.error(createConnectionError(channelName));
+            } 
+          });
     }
-    public void onContinueRequest(@NonNull TimerRequest requestArg, @NonNull Reply<Void> callback) {
+    public void onContinueRequest(@NonNull TimerRequest requestArg, @NonNull VoidResult result) {
+      final String channelName = "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onContinueRequest";
       BasicMessageChannel<Object> channel =
           new BasicMessageChannel<>(
-              binaryMessenger, "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onContinueRequest", getCodec());
+              binaryMessenger, channelName, getCodec());
       channel.send(
           new ArrayList<Object>(Collections.singletonList(requestArg)),
-          channelReply -> callback.reply(null));
+          channelReply -> {
+            if (channelReply instanceof List) {
+              List<Object> listReply = (List<Object>) channelReply;
+              if (listReply.size() > 1) {
+                result.error(new FlutterError((String) listReply.get(0), (String) listReply.get(1), (String) listReply.get(2)));
+              } else {
+                result.success();
+              }
+            }  else {
+              result.error(createConnectionError(channelName));
+            } 
+          });
     }
-    public void onPauseRequest(@NonNull TimerRequest requestArg, @NonNull Reply<Void> callback) {
+    public void onPauseRequest(@NonNull TimerRequest requestArg, @NonNull VoidResult result) {
+      final String channelName = "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onPauseRequest";
       BasicMessageChannel<Object> channel =
           new BasicMessageChannel<>(
-              binaryMessenger, "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onPauseRequest", getCodec());
+              binaryMessenger, channelName, getCodec());
       channel.send(
           new ArrayList<Object>(Collections.singletonList(requestArg)),
-          channelReply -> callback.reply(null));
+          channelReply -> {
+            if (channelReply instanceof List) {
+              List<Object> listReply = (List<Object>) channelReply;
+              if (listReply.size() > 1) {
+                result.error(new FlutterError((String) listReply.get(0), (String) listReply.get(1), (String) listReply.get(2)));
+              } else {
+                result.success();
+              }
+            }  else {
+              result.error(createConnectionError(channelName));
+            } 
+          });
     }
-    public void onCancelRequest(@NonNull TimerRequest requestArg, @NonNull Reply<Void> callback) {
+    public void onCancelRequest(@NonNull TimerRequest requestArg, @NonNull VoidResult result) {
+      final String channelName = "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onCancelRequest";
       BasicMessageChannel<Object> channel =
           new BasicMessageChannel<>(
-              binaryMessenger, "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onCancelRequest", getCodec());
+              binaryMessenger, channelName, getCodec());
       channel.send(
           new ArrayList<Object>(Collections.singletonList(requestArg)),
-          channelReply -> callback.reply(null));
+          channelReply -> {
+            if (channelReply instanceof List) {
+              List<Object> listReply = (List<Object>) channelReply;
+              if (listReply.size() > 1) {
+                result.error(new FlutterError((String) listReply.get(0), (String) listReply.get(1), (String) listReply.get(2)));
+              } else {
+                result.success();
+              }
+            }  else {
+              result.error(createConnectionError(channelName));
+            } 
+          });
     }
-    public void onRestartRequest(@NonNull TimerRequest requestArg, @NonNull Reply<Void> callback) {
+    public void onRestartRequest(@NonNull TimerRequest requestArg, @NonNull VoidResult result) {
+      final String channelName = "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onRestartRequest";
       BasicMessageChannel<Object> channel =
           new BasicMessageChannel<>(
-              binaryMessenger, "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onRestartRequest", getCodec());
+              binaryMessenger, channelName, getCodec());
       channel.send(
           new ArrayList<Object>(Collections.singletonList(requestArg)),
-          channelReply -> callback.reply(null));
+          channelReply -> {
+            if (channelReply instanceof List) {
+              List<Object> listReply = (List<Object>) channelReply;
+              if (listReply.size() > 1) {
+                result.error(new FlutterError((String) listReply.get(0), (String) listReply.get(1), (String) listReply.get(2)));
+              } else {
+                result.success();
+              }
+            }  else {
+              result.error(createConnectionError(channelName));
+            } 
+          });
     }
-    public void onOpen(@NonNull OpenRequest requestArg, @NonNull Reply<Void> callback) {
+    public void onOpen(@NonNull OpenRequest requestArg, @NonNull VoidResult result) {
+      final String channelName = "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onOpen";
       BasicMessageChannel<Object> channel =
           new BasicMessageChannel<>(
-              binaryMessenger, "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onOpen", getCodec());
+              binaryMessenger, channelName, getCodec());
       channel.send(
           new ArrayList<Object>(Collections.singletonList(requestArg)),
-          channelReply -> callback.reply(null));
+          channelReply -> {
+            if (channelReply instanceof List) {
+              List<Object> listReply = (List<Object>) channelReply;
+              if (listReply.size() > 1) {
+                result.error(new FlutterError((String) listReply.get(0), (String) listReply.get(1), (String) listReply.get(2)));
+              } else {
+                result.success();
+              }
+            }  else {
+              result.error(createConnectionError(channelName));
+            } 
+          });
     }
-    public void onAlarm(@NonNull TimerRequest requestArg, @NonNull Reply<Void> callback) {
+    public void onAlarm(@NonNull TimerRequest requestArg, @NonNull VoidResult result) {
+      final String channelName = "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onAlarm";
       BasicMessageChannel<Object> channel =
           new BasicMessageChannel<>(
-              binaryMessenger, "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onAlarm", getCodec());
+              binaryMessenger, channelName, getCodec());
       channel.send(
           new ArrayList<Object>(Collections.singletonList(requestArg)),
-          channelReply -> callback.reply(null));
+          channelReply -> {
+            if (channelReply instanceof List) {
+              List<Object> listReply = (List<Object>) channelReply;
+              if (listReply.size() > 1) {
+                result.error(new FlutterError((String) listReply.get(0), (String) listReply.get(1), (String) listReply.get(2)));
+              } else {
+                result.success();
+              }
+            }  else {
+              result.error(createConnectionError(channelName));
+            } 
+          });
     }
-    public void onWidgetUpdate(@NonNull Reply<WidgetUpdateResponse> callback) {
+    public void onWidgetUpdate(@NonNull Result<WidgetUpdateResponse> result) {
+      final String channelName = "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onWidgetUpdate";
       BasicMessageChannel<Object> channel =
           new BasicMessageChannel<>(
-              binaryMessenger, "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onWidgetUpdate", getCodec());
+              binaryMessenger, channelName, getCodec());
       channel.send(
           null,
           channelReply -> {
-            @SuppressWarnings("ConstantConditions")
-            WidgetUpdateResponse output = (WidgetUpdateResponse) channelReply;
-            callback.reply(output);
+            if (channelReply instanceof List) {
+              List<Object> listReply = (List<Object>) channelReply;
+              if (listReply.size() > 1) {
+                result.error(new FlutterError((String) listReply.get(0), (String) listReply.get(1), (String) listReply.get(2)));
+              } else if (listReply.get(0) == null) {
+                result.error(new FlutterError("null-error", "Flutter api returned null value for non-null return value.", ""));
+              } else {
+                @SuppressWarnings("ConstantConditions")
+                WidgetUpdateResponse output = (WidgetUpdateResponse) listReply.get(0);
+                result.success(output);
+              }
+            }  else {
+              result.error(createConnectionError(channelName));
+            } 
           });
     }
-    public void onWidgetStartTimer(@NonNull Reply<Void> callback) {
+    public void onWidgetStartTimer(@NonNull VoidResult result) {
+      final String channelName = "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onWidgetStartTimer";
       BasicMessageChannel<Object> channel =
           new BasicMessageChannel<>(
-              binaryMessenger, "dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onWidgetStartTimer", getCodec());
+              binaryMessenger, channelName, getCodec());
       channel.send(
           null,
-          channelReply -> callback.reply(null));
+          channelReply -> {
+            if (channelReply instanceof List) {
+              List<Object> listReply = (List<Object>) channelReply;
+              if (listReply.size() > 1) {
+                result.error(new FlutterError((String) listReply.get(0), (String) listReply.get(1), (String) listReply.get(2)));
+              } else {
+                result.success();
+              }
+            }  else {
+              result.error(createConnectionError(channelName));
+            } 
+          });
     }
   }
 }
