@@ -1,12 +1,29 @@
-// Autogenerated from Pigeon (v10.1.6), do not edit directly.
+// Autogenerated from Pigeon (v16.0.0), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
-// ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import, unnecessary_parenthesis, prefer_null_aware_operators, omit_local_variable_types, unused_shown_name, unnecessary_import
+// ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import, unnecessary_parenthesis, prefer_null_aware_operators, omit_local_variable_types, unused_shown_name, unnecessary_import, no_leading_underscores_for_local_identifiers
 
 import 'dart:async';
 import 'dart:typed_data' show Float64List, Int32List, Int64List, Uint8List;
 
 import 'package:flutter/foundation.dart' show ReadBuffer, WriteBuffer;
 import 'package:flutter/services.dart';
+
+PlatformException _createConnectionError(String channelName) {
+  return PlatformException(
+    code: 'channel-error',
+    message: 'Unable to establish connection on channel: "$channelName".',
+  );
+}
+
+List<Object?> wrapResponse({Object? result, PlatformException? error, bool empty = false}) {
+  if (empty) {
+    return <Object?>[];
+  }
+  if (error == null) {
+    return <Object?>[result];
+  }
+  return <Object?>[error.code, error.message, error.details];
+}
 
 class NotificationRequest {
   NotificationRequest({
@@ -576,135 +593,135 @@ class HostTimerApi {
   /// available for dependency injection.  If it is left null, the default
   /// BinaryMessenger will be used which routes to the host platform.
   HostTimerApi({BinaryMessenger? binaryMessenger})
-      : _binaryMessenger = binaryMessenger;
-  final BinaryMessenger? _binaryMessenger;
+      : __pigeon_binaryMessenger = binaryMessenger;
+  final BinaryMessenger? __pigeon_binaryMessenger;
 
-  static const MessageCodec<Object?> codec = _HostTimerApiCodec();
+  static const MessageCodec<Object?> pigeonChannelCodec = _HostTimerApiCodec();
 
-  Future<NotificationResponse> showRunningNotification(RunningNotificationRequest arg_request) async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.sleep_timer.HostTimerApi.showRunningNotification', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(<Object?>[arg_request]) as List<Object?>?;
-    if (replyList == null) {
+  Future<NotificationResponse> showRunningNotification(RunningNotificationRequest request) async {
+    const String __pigeon_channelName = 'dev.flutter.pigeon.sleep_timer.HostTimerApi.showRunningNotification';
+    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+      __pigeon_channelName,
+      pigeonChannelCodec,
+      binaryMessenger: __pigeon_binaryMessenger,
+    );
+    final List<Object?>? __pigeon_replyList =
+        await __pigeon_channel.send(<Object?>[request]) as List<Object?>?;
+    if (__pigeon_replyList == null) {
+      throw _createConnectionError(__pigeon_channelName);
+    } else if (__pigeon_replyList.length > 1) {
       throw PlatformException(
-        code: 'channel-error',
-        message: 'Unable to establish connection on channel.',
+        code: __pigeon_replyList[0]! as String,
+        message: __pigeon_replyList[1] as String?,
+        details: __pigeon_replyList[2],
       );
-    } else if (replyList.length > 1) {
-      throw PlatformException(
-        code: replyList[0]! as String,
-        message: replyList[1] as String?,
-        details: replyList[2],
-      );
-    } else if (replyList[0] == null) {
+    } else if (__pigeon_replyList[0] == null) {
       throw PlatformException(
         code: 'null-error',
         message: 'Host platform returned null value for non-null return value.',
       );
     } else {
-      return (replyList[0] as NotificationResponse?)!;
+      return (__pigeon_replyList[0] as NotificationResponse?)!;
     }
   }
 
-  Future<NotificationResponse> showPausingNotification(TimeNotificationRequest arg_request) async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.sleep_timer.HostTimerApi.showPausingNotification', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(<Object?>[arg_request]) as List<Object?>?;
-    if (replyList == null) {
+  Future<NotificationResponse> showPausingNotification(TimeNotificationRequest request) async {
+    const String __pigeon_channelName = 'dev.flutter.pigeon.sleep_timer.HostTimerApi.showPausingNotification';
+    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+      __pigeon_channelName,
+      pigeonChannelCodec,
+      binaryMessenger: __pigeon_binaryMessenger,
+    );
+    final List<Object?>? __pigeon_replyList =
+        await __pigeon_channel.send(<Object?>[request]) as List<Object?>?;
+    if (__pigeon_replyList == null) {
+      throw _createConnectionError(__pigeon_channelName);
+    } else if (__pigeon_replyList.length > 1) {
       throw PlatformException(
-        code: 'channel-error',
-        message: 'Unable to establish connection on channel.',
+        code: __pigeon_replyList[0]! as String,
+        message: __pigeon_replyList[1] as String?,
+        details: __pigeon_replyList[2],
       );
-    } else if (replyList.length > 1) {
-      throw PlatformException(
-        code: replyList[0]! as String,
-        message: replyList[1] as String?,
-        details: replyList[2],
-      );
-    } else if (replyList[0] == null) {
+    } else if (__pigeon_replyList[0] == null) {
       throw PlatformException(
         code: 'null-error',
         message: 'Host platform returned null value for non-null return value.',
       );
     } else {
-      return (replyList[0] as NotificationResponse?)!;
+      return (__pigeon_replyList[0] as NotificationResponse?)!;
     }
   }
 
-  Future<NotificationResponse> showElapsedNotification(NotificationRequest arg_request) async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.sleep_timer.HostTimerApi.showElapsedNotification', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(<Object?>[arg_request]) as List<Object?>?;
-    if (replyList == null) {
+  Future<NotificationResponse> showElapsedNotification(NotificationRequest request) async {
+    const String __pigeon_channelName = 'dev.flutter.pigeon.sleep_timer.HostTimerApi.showElapsedNotification';
+    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+      __pigeon_channelName,
+      pigeonChannelCodec,
+      binaryMessenger: __pigeon_binaryMessenger,
+    );
+    final List<Object?>? __pigeon_replyList =
+        await __pigeon_channel.send(<Object?>[request]) as List<Object?>?;
+    if (__pigeon_replyList == null) {
+      throw _createConnectionError(__pigeon_channelName);
+    } else if (__pigeon_replyList.length > 1) {
       throw PlatformException(
-        code: 'channel-error',
-        message: 'Unable to establish connection on channel.',
+        code: __pigeon_replyList[0]! as String,
+        message: __pigeon_replyList[1] as String?,
+        details: __pigeon_replyList[2],
       );
-    } else if (replyList.length > 1) {
-      throw PlatformException(
-        code: replyList[0]! as String,
-        message: replyList[1] as String?,
-        details: replyList[2],
-      );
-    } else if (replyList[0] == null) {
+    } else if (__pigeon_replyList[0] == null) {
       throw PlatformException(
         code: 'null-error',
         message: 'Host platform returned null value for non-null return value.',
       );
     } else {
-      return (replyList[0] as NotificationResponse?)!;
+      return (__pigeon_replyList[0] as NotificationResponse?)!;
     }
   }
 
-  Future<CancelResponse> cancelTimer(CancelRequest arg_request) async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.sleep_timer.HostTimerApi.cancelTimer', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(<Object?>[arg_request]) as List<Object?>?;
-    if (replyList == null) {
+  Future<CancelResponse> cancelTimer(CancelRequest request) async {
+    const String __pigeon_channelName = 'dev.flutter.pigeon.sleep_timer.HostTimerApi.cancelTimer';
+    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+      __pigeon_channelName,
+      pigeonChannelCodec,
+      binaryMessenger: __pigeon_binaryMessenger,
+    );
+    final List<Object?>? __pigeon_replyList =
+        await __pigeon_channel.send(<Object?>[request]) as List<Object?>?;
+    if (__pigeon_replyList == null) {
+      throw _createConnectionError(__pigeon_channelName);
+    } else if (__pigeon_replyList.length > 1) {
       throw PlatformException(
-        code: 'channel-error',
-        message: 'Unable to establish connection on channel.',
+        code: __pigeon_replyList[0]! as String,
+        message: __pigeon_replyList[1] as String?,
+        details: __pigeon_replyList[2],
       );
-    } else if (replyList.length > 1) {
-      throw PlatformException(
-        code: replyList[0]! as String,
-        message: replyList[1] as String?,
-        details: replyList[2],
-      );
-    } else if (replyList[0] == null) {
+    } else if (__pigeon_replyList[0] == null) {
       throw PlatformException(
         code: 'null-error',
         message: 'Host platform returned null value for non-null return value.',
       );
     } else {
-      return (replyList[0] as CancelResponse?)!;
+      return (__pigeon_replyList[0] as CancelResponse?)!;
     }
   }
 
-  Future<void> toggleExtendByShake(ToggleRequest arg_request) async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.sleep_timer.HostTimerApi.toggleExtendByShake', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(<Object?>[arg_request]) as List<Object?>?;
-    if (replyList == null) {
+  Future<void> toggleExtendByShake(ToggleRequest request) async {
+    const String __pigeon_channelName = 'dev.flutter.pigeon.sleep_timer.HostTimerApi.toggleExtendByShake';
+    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+      __pigeon_channelName,
+      pigeonChannelCodec,
+      binaryMessenger: __pigeon_binaryMessenger,
+    );
+    final List<Object?>? __pigeon_replyList =
+        await __pigeon_channel.send(<Object?>[request]) as List<Object?>?;
+    if (__pigeon_replyList == null) {
+      throw _createConnectionError(__pigeon_channelName);
+    } else if (__pigeon_replyList.length > 1) {
       throw PlatformException(
-        code: 'channel-error',
-        message: 'Unable to establish connection on channel.',
-      );
-    } else if (replyList.length > 1) {
-      throw PlatformException(
-        code: replyList[0]! as String,
-        message: replyList[1] as String?,
-        details: replyList[2],
+        code: __pigeon_replyList[0]! as String,
+        message: __pigeon_replyList[1] as String?,
+        details: __pigeon_replyList[2],
       );
     } else {
       return;
@@ -712,75 +729,75 @@ class HostTimerApi {
   }
 
   Future<InstalledAppsResponse> getInstalledPlayerApps() async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.sleep_timer.HostTimerApi.getInstalledPlayerApps', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(null) as List<Object?>?;
-    if (replyList == null) {
+    const String __pigeon_channelName = 'dev.flutter.pigeon.sleep_timer.HostTimerApi.getInstalledPlayerApps';
+    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+      __pigeon_channelName,
+      pigeonChannelCodec,
+      binaryMessenger: __pigeon_binaryMessenger,
+    );
+    final List<Object?>? __pigeon_replyList =
+        await __pigeon_channel.send(null) as List<Object?>?;
+    if (__pigeon_replyList == null) {
+      throw _createConnectionError(__pigeon_channelName);
+    } else if (__pigeon_replyList.length > 1) {
       throw PlatformException(
-        code: 'channel-error',
-        message: 'Unable to establish connection on channel.',
+        code: __pigeon_replyList[0]! as String,
+        message: __pigeon_replyList[1] as String?,
+        details: __pigeon_replyList[2],
       );
-    } else if (replyList.length > 1) {
-      throw PlatformException(
-        code: replyList[0]! as String,
-        message: replyList[1] as String?,
-        details: replyList[2],
-      );
-    } else if (replyList[0] == null) {
+    } else if (__pigeon_replyList[0] == null) {
       throw PlatformException(
         code: 'null-error',
         message: 'Host platform returned null value for non-null return value.',
       );
     } else {
-      return (replyList[0] as InstalledAppsResponse?)!;
+      return (__pigeon_replyList[0] as InstalledAppsResponse?)!;
     }
   }
 
   Future<InstalledAppsResponse> getInstalledAlarmApps() async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.sleep_timer.HostTimerApi.getInstalledAlarmApps', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(null) as List<Object?>?;
-    if (replyList == null) {
+    const String __pigeon_channelName = 'dev.flutter.pigeon.sleep_timer.HostTimerApi.getInstalledAlarmApps';
+    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+      __pigeon_channelName,
+      pigeonChannelCodec,
+      binaryMessenger: __pigeon_binaryMessenger,
+    );
+    final List<Object?>? __pigeon_replyList =
+        await __pigeon_channel.send(null) as List<Object?>?;
+    if (__pigeon_replyList == null) {
+      throw _createConnectionError(__pigeon_channelName);
+    } else if (__pigeon_replyList.length > 1) {
       throw PlatformException(
-        code: 'channel-error',
-        message: 'Unable to establish connection on channel.',
+        code: __pigeon_replyList[0]! as String,
+        message: __pigeon_replyList[1] as String?,
+        details: __pigeon_replyList[2],
       );
-    } else if (replyList.length > 1) {
-      throw PlatformException(
-        code: replyList[0]! as String,
-        message: replyList[1] as String?,
-        details: replyList[2],
-      );
-    } else if (replyList[0] == null) {
+    } else if (__pigeon_replyList[0] == null) {
       throw PlatformException(
         code: 'null-error',
         message: 'Host platform returned null value for non-null return value.',
       );
     } else {
-      return (replyList[0] as InstalledAppsResponse?)!;
+      return (__pigeon_replyList[0] as InstalledAppsResponse?)!;
     }
   }
 
-  Future<void> launchApp(LaunchAppRequest arg_request) async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.sleep_timer.HostTimerApi.launchApp', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(<Object?>[arg_request]) as List<Object?>?;
-    if (replyList == null) {
+  Future<void> launchApp(LaunchAppRequest request) async {
+    const String __pigeon_channelName = 'dev.flutter.pigeon.sleep_timer.HostTimerApi.launchApp';
+    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+      __pigeon_channelName,
+      pigeonChannelCodec,
+      binaryMessenger: __pigeon_binaryMessenger,
+    );
+    final List<Object?>? __pigeon_replyList =
+        await __pigeon_channel.send(<Object?>[request]) as List<Object?>?;
+    if (__pigeon_replyList == null) {
+      throw _createConnectionError(__pigeon_channelName);
+    } else if (__pigeon_replyList.length > 1) {
       throw PlatformException(
-        code: 'channel-error',
-        message: 'Unable to establish connection on channel.',
-      );
-    } else if (replyList.length > 1) {
-      throw PlatformException(
-        code: replyList[0]! as String,
-        message: replyList[1] as String?,
-        details: replyList[2],
+        code: __pigeon_replyList[0]! as String,
+        message: __pigeon_replyList[1] as String?,
+        details: __pigeon_replyList[2],
       );
     } else {
       return;
@@ -832,7 +849,7 @@ class _FlutterTimerApiCodec extends StandardMessageCodec {
 }
 
 abstract class FlutterTimerApi {
-  static const MessageCodec<Object?> codec = _FlutterTimerApiCodec();
+  static const MessageCodec<Object?> pigeonChannelCodec = _FlutterTimerApiCodec();
 
   void onExtendTime(ExtendTimeRequest request);
 
@@ -856,182 +873,240 @@ abstract class FlutterTimerApi {
 
   static void setup(FlutterTimerApi? api, {BinaryMessenger? binaryMessenger}) {
     {
-      final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onExtendTime', codec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onExtendTime', pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
-        channel.setMessageHandler(null);
+        __pigeon_channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object? message) async {
+        __pigeon_channel.setMessageHandler((Object? message) async {
           assert(message != null,
           'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onExtendTime was null.');
           final List<Object?> args = (message as List<Object?>?)!;
           final ExtendTimeRequest? arg_request = (args[0] as ExtendTimeRequest?);
           assert(arg_request != null,
               'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onExtendTime was null, expected non-null ExtendTimeRequest.');
-          api.onExtendTime(arg_request!);
-          return;
+          try {
+            api.onExtendTime(arg_request!);
+            return wrapResponse(empty: true);
+          } on PlatformException catch (e) {
+            return wrapResponse(error: e);
+          }          catch (e) {
+            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onCountDown', codec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onCountDown', pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
-        channel.setMessageHandler(null);
+        __pigeon_channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object? message) async {
+        __pigeon_channel.setMessageHandler((Object? message) async {
           assert(message != null,
           'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onCountDown was null.');
           final List<Object?> args = (message as List<Object?>?)!;
           final CountDownRequest? arg_request = (args[0] as CountDownRequest?);
           assert(arg_request != null,
               'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onCountDown was null, expected non-null CountDownRequest.');
-          api.onCountDown(arg_request!);
-          return;
+          try {
+            api.onCountDown(arg_request!);
+            return wrapResponse(empty: true);
+          } on PlatformException catch (e) {
+            return wrapResponse(error: e);
+          }          catch (e) {
+            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onContinueRequest', codec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onContinueRequest', pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
-        channel.setMessageHandler(null);
+        __pigeon_channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object? message) async {
+        __pigeon_channel.setMessageHandler((Object? message) async {
           assert(message != null,
           'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onContinueRequest was null.');
           final List<Object?> args = (message as List<Object?>?)!;
           final TimerRequest? arg_request = (args[0] as TimerRequest?);
           assert(arg_request != null,
               'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onContinueRequest was null, expected non-null TimerRequest.');
-          api.onContinueRequest(arg_request!);
-          return;
+          try {
+            api.onContinueRequest(arg_request!);
+            return wrapResponse(empty: true);
+          } on PlatformException catch (e) {
+            return wrapResponse(error: e);
+          }          catch (e) {
+            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onPauseRequest', codec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onPauseRequest', pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
-        channel.setMessageHandler(null);
+        __pigeon_channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object? message) async {
+        __pigeon_channel.setMessageHandler((Object? message) async {
           assert(message != null,
           'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onPauseRequest was null.');
           final List<Object?> args = (message as List<Object?>?)!;
           final TimerRequest? arg_request = (args[0] as TimerRequest?);
           assert(arg_request != null,
               'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onPauseRequest was null, expected non-null TimerRequest.');
-          api.onPauseRequest(arg_request!);
-          return;
+          try {
+            api.onPauseRequest(arg_request!);
+            return wrapResponse(empty: true);
+          } on PlatformException catch (e) {
+            return wrapResponse(error: e);
+          }          catch (e) {
+            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onCancelRequest', codec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onCancelRequest', pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
-        channel.setMessageHandler(null);
+        __pigeon_channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object? message) async {
+        __pigeon_channel.setMessageHandler((Object? message) async {
           assert(message != null,
           'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onCancelRequest was null.');
           final List<Object?> args = (message as List<Object?>?)!;
           final TimerRequest? arg_request = (args[0] as TimerRequest?);
           assert(arg_request != null,
               'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onCancelRequest was null, expected non-null TimerRequest.');
-          api.onCancelRequest(arg_request!);
-          return;
+          try {
+            api.onCancelRequest(arg_request!);
+            return wrapResponse(empty: true);
+          } on PlatformException catch (e) {
+            return wrapResponse(error: e);
+          }          catch (e) {
+            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onRestartRequest', codec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onRestartRequest', pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
-        channel.setMessageHandler(null);
+        __pigeon_channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object? message) async {
+        __pigeon_channel.setMessageHandler((Object? message) async {
           assert(message != null,
           'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onRestartRequest was null.');
           final List<Object?> args = (message as List<Object?>?)!;
           final TimerRequest? arg_request = (args[0] as TimerRequest?);
           assert(arg_request != null,
               'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onRestartRequest was null, expected non-null TimerRequest.');
-          api.onRestartRequest(arg_request!);
-          return;
+          try {
+            api.onRestartRequest(arg_request!);
+            return wrapResponse(empty: true);
+          } on PlatformException catch (e) {
+            return wrapResponse(error: e);
+          }          catch (e) {
+            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onOpen', codec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onOpen', pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
-        channel.setMessageHandler(null);
+        __pigeon_channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object? message) async {
+        __pigeon_channel.setMessageHandler((Object? message) async {
           assert(message != null,
           'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onOpen was null.');
           final List<Object?> args = (message as List<Object?>?)!;
           final OpenRequest? arg_request = (args[0] as OpenRequest?);
           assert(arg_request != null,
               'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onOpen was null, expected non-null OpenRequest.');
-          api.onOpen(arg_request!);
-          return;
+          try {
+            api.onOpen(arg_request!);
+            return wrapResponse(empty: true);
+          } on PlatformException catch (e) {
+            return wrapResponse(error: e);
+          }          catch (e) {
+            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onAlarm', codec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onAlarm', pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
-        channel.setMessageHandler(null);
+        __pigeon_channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object? message) async {
+        __pigeon_channel.setMessageHandler((Object? message) async {
           assert(message != null,
           'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onAlarm was null.');
           final List<Object?> args = (message as List<Object?>?)!;
           final TimerRequest? arg_request = (args[0] as TimerRequest?);
           assert(arg_request != null,
               'Argument for dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onAlarm was null, expected non-null TimerRequest.');
-          api.onAlarm(arg_request!);
-          return;
+          try {
+            api.onAlarm(arg_request!);
+            return wrapResponse(empty: true);
+          } on PlatformException catch (e) {
+            return wrapResponse(error: e);
+          }          catch (e) {
+            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onWidgetUpdate', codec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onWidgetUpdate', pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
-        channel.setMessageHandler(null);
+        __pigeon_channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object? message) async {
-          // ignore message
-          final WidgetUpdateResponse output = api.onWidgetUpdate();
-          return output;
+        __pigeon_channel.setMessageHandler((Object? message) async {
+          try {
+            final WidgetUpdateResponse output = api.onWidgetUpdate();
+            return wrapResponse(result: output);
+          } on PlatformException catch (e) {
+            return wrapResponse(error: e);
+          }          catch (e) {
+            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          }
         });
       }
     }
     {
-      final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onWidgetStartTimer', codec,
+      final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.sleep_timer.FlutterTimerApi.onWidgetStartTimer', pigeonChannelCodec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
-        channel.setMessageHandler(null);
+        __pigeon_channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object? message) async {
-          // ignore message
-          api.onWidgetStartTimer();
-          return;
+        __pigeon_channel.setMessageHandler((Object? message) async {
+          try {
+            api.onWidgetStartTimer();
+            return wrapResponse(empty: true);
+          } on PlatformException catch (e) {
+            return wrapResponse(error: e);
+          }          catch (e) {
+            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          }
         });
       }
     }
